More benchmarks

+++ train_programs/read_context_switches.h
@@ -5,10 +5,9 @@ extern "C" {
 #ifndef __READ_CONTEXT_SWITCHES_H
 #define __READ_CONTEXT_SWITCHES_H
 
-int read_context_switches(bool isVoluntary, pid_t tid);
+long long read_context_switches(bool isVoluntary, pid_t tid);
 
 #endif
-
 
 #ifdef __cplusplus
 }
